Question_4/requirements.txt
numpy==1.26.3
fastapi==0.111.1
pytest==8.1.1
pillow==10.2.0
uvicorn==0.30.3